refactor: get file stats function and apply requested change
